Fixed planet deletion
Planet deletion has been working correctly, but you were able to delete all planets of an account, which renders the account invalid.

This is fixed by redirecting the user to account deletion confirmation. Even better would be to provide the user with the possibility to create a new planet on this page.